Movie - Load Fanart tab images in multiple backgroundworkers. Improve loading of preview images.

+++ Utilites/DownloadCache.vb
@@ -161,20 +161,11 @@ Public Class DownloadCache
                     End If
                     Return True
                 End If
-                If ForceDownload AndAlso File.Exists(CachePath) Then
-                    File.Delete(CachePath)
-                End If
+
+                If ForceDownload AndAlso File.Exists(CachePath) Then File.Delete(CachePath)
 
                 Try
                     Dim webReq As HttpWebRequest = DirectCast(WebRequest.Create(URL), HttpWebRequest)
-                    'If Utilities.MCProxy.Item(0).ToLower = "false" Then
-                    '   ' Dim myProxy As New WebProxy("myproxy", 80)
-                    '    webReq.Proxy = Nothing
-                    'Else
-                    '    Dim myProxy As New WebProxy(Utilities.MCProxy.Item(1), Convert.ToInt32(Utilities.MCProxy.Item(2)))
-                    '    myProxy.Credentials = New NetworkCredential(Utilities.MCProxy.Item(3), Utilities.MCProxy.item(4))
-                    '    webReq.Proxy = myProxy
-                    'End If
                     webReq.Proxy = Utilities.MyProxy
                     webReq.AllowAutoRedirect = True
                     webReq.AutomaticDecompression = DecompressionMethods.GZip Or DecompressionMethods.Deflate
@@ -235,14 +226,6 @@ Public Class DownloadCache
                 End If
                 Try
                     Dim webReq As HttpWebRequest = DirectCast(WebRequest.Create(URL), HttpWebRequest)
-                    'If Utilities.MCProxy.Item(0).ToLower = "false" Then
-                    '   ' Dim myProxy As New WebProxy("myproxy", 80)
-                    '    webReq.Proxy = Nothing
-                    'Else
-                    '    Dim myProxy As New WebProxy(Utilities.MCProxy.Item(1), Convert.ToInt32(Utilities.MCProxy.Item(2)))
-                    '    myProxy.Credentials = New NetworkCredential(Utilities.MCProxy.Item(3), Utilities.MCProxy.item(4))
-                    '    webReq.Proxy = myProxy
-                    'End If
                     webReq.Proxy = Utilities.MyProxy
                     webReq.AllowAutoRedirect = True
                     webReq.AutomaticDecompression = DecompressionMethods.GZip Or DecompressionMethods.Deflate
